chore: rename config/ts-node to config/ts-path
