+++ src/config/index.ts
@@ -1,3 +1,3 @@
 export {Config} from './config'
 export {Plugin} from './plugin'
-export {tsPath} from './ts-node'
+export {tsPath} from './ts-path'
