+++ src/config/plugin.ts
@@ -14,7 +14,7 @@ import {cacheCommand} from '../util/cache-command'
 import {findRoot} from '../util/find-root'
 import {readJson, requireJson} from '../util/fs'
 import {castArray, compact} from '../util/util'
-import {tsPath} from './ts-node'
+import {tsPath} from './ts-path'
 import {Debug, getCommandIdPermutations} from './util'
 
 const _pjson = requireJson<PJSON>(__dirname, '..', '..', 'package.json')
